Basic enemy shooting collision logic.

+++ src/org/prototype/projectiles/Projectile.java
@@ -7,11 +7,11 @@ import org.newdawn.slick.geom.Shape;
 public abstract class Projectile {
 	protected int velocity;
 	protected int team;
-	public int x;
-	public int y;
+	public float x;
+	public float y;
 	public Shape boundingBox;
 	
-	public Projectile (int newVelocity, int newTeam, int startingX, int startingY, int width, int height) {
+	public Projectile (int newVelocity, int newTeam, float startingX, float startingY, int width, int height) {
 		velocity = newVelocity;
 		team = newTeam;
 		x = startingX;
